Allow building with the homebrew version of openssl
This change allows you to install with the --brewed-openssl option to
build and use the current 1.0.1c version of openssl that is available
in the main Homebrew repository.

We need this because of immediate segfaulting of PHP when connecting to
a NTLM protected .NET webservice over HTTPS. This is solved when
updating to this openssl version.

We figured this out when it al worked just fine on a recent Ubuntu, and
on our OSX dev environments when we used http, but when switching to
https the segfaults started.