Specifying right phpize for brewed PHP
Installing extensions when the phpize in $PATH is incorrect for that
extension will result in an InvalidPhpizeError. Those without PHP in
their $PATH at all will recieve a UnsupportedPhpApiError, since we
cannot detect `phpize`.

Currently supported are PHP53 and PHP54. Support for other versions
should update line 34 of abstract-php-extension.rb to include the
PHP API Version.

When installing against a Homebrew version of PHP that is not
currently linked, it is possible to circumvent the path using the
`--with-homebrew-php` argument.

Note that calling `system "phpize"` will result in not using the
proper phpize. Instead, `safe_phpize` should be used instead.

Closes #113
Closes #156
Closes #173