Remove all pear.conf files from php path
Having a `pear.conf` file appears to block the `./configure` or
`make` step of PHP from writing a `pear` binary, resulting in a
failure to install because `brew install` later calls
`pear config-set`.

We should take a backup of the file and move it back in case of a
failed compile.

Thanks to Andy Blyler, David Rekowski, Todd Bloom, and Mark
Shropshire - and everyone else! - for digging into the issue and
confirming the fix.

Implements #101
Closes #124
Closes #125
Closes #191
Closes #206
Closes #209
Closes #223
Closes #224