Moved logic for installing a php version to abstract-php
This should allow us to more easily support multiple
PHP versions with a single formula. Note that we may
opt to remove old extensions, or not include extensions,
for a given PHP version.

One other caveat is that not all PHP versions will have
the same options compiled in, even with this standard base.