No longer build with inline intl support
As of mxcl/homebrew#03ed757c, icu4c somehow breaks when being built
with php due to undefined symbols. If being built as an extension,
everything works fine.

Rather than deal with people asking why it doesn't work, I am forcing
the installation of of intl as an extension. A warning will be raised
if the intl flag is passed.

Unfortunately, a post-install dependency isn't straightforward, so at
this time this is the best we can do.

If icu4c becomes compliant with PHP building in the future, we can
revert this change and allow users to rebuild with inline intl support.

Closes #378
Closes #383
Closes #388
Closes #389