Build php-cs-fixer.phar on the fly. Closes #377
@fabpot refuses to release tagged builds for php-cs-fixer because it is
pre-alpha software. As such, it is impossible to follow his daily rebuilds
of the released phar, making it a chore to maintain php-cs-fixer.

This commit retools the work by @davidneimeyer in his php-cs-fixer_installer
repo into a homebrew formulae, allowing us to build the phar file on the fly.
Because of the phar requirements, phar.readonly must be set to off for This
formulae to run properly.

Note that detection of installed php bins may fail if you are using HEAD for
your homebrew-php install.