Fixes for OS 10.5.x Leopard PHP build errors. PHP Makefile always puts system libraries first
Due to a bug upstream with PHP 5.x (https://bugs.php.net/bug.php?id=44294),
the system libraries are always on the build search path before library
paths passed in the by the user. This causes builds to fail on OS 10.5
Leopard, where many system libraries are out-of-date.

Homebrew OpenSSL required to build PHP5.4 on 10.5.

Due to system OpenSSL being too far out-of-date, PHP 5.4 will not build
without --with-homebrew-openssl installation argument. This commit
defaults that option if MacOS.version == :leopard, otherwise users of
newer versions of MacOS are free to use system or Homebrew OpenSSL as
they desire.